Create containers and move needed components
